+++ NEWS.md
@@ -1,3 +1,7 @@
+- New method `to_int` for quick conversion of values to `Int`.
+
+- `bcastdim` and `bcastdims` have been deprecated in favor of `bcastsize`.
+
 - `checkdimensions` has been deprecated in favor of `check_dimensions`.
 
 - Rubber index operator `…` has been renamed as `..` (which is easier to type) as in